Intervals and Harmonics
Today, we got to play a bit with with harmonics.
Yes, do, re, mi, fa, so, la, ti, do is definitely a way to practice hearing the pitches. 
Write and journal a few measures that really stuck through. @meepmeepme @Nathan090 @coolernow
Undertale. I'll have to find time for that one! :)